Working login system for both
Have fixed the login. Can now login with either admin or orchard admin credentials. Spent a day working on this as it kept referring back to the circa home page and wouldn't recognize the orchard credentials, only the admin. This is now working and I plan to work on the add employee function now so that the admin can add and manage the employees of the orchard.